% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miniphylopic-package.R
\docType{package}
\name{miniphylopic-package}
\alias{miniphylopic}
\alias{miniphylopic-package}
\title{miniphylopic: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Russell Dinnage \email{r.dinnage@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
